High dimensional mediation analysis.

High dimensional mediation analysis with confounding adjustment. Methods of confounding adjustment include regression adjustment, propensity score adjustment, inverse probability weighting, overlapping weighting.

Usage

himaconf(
  X,
  M,
  Y,
  COV,
  k = 2,
  CONFmethod = c("RA", "PS", "IPW", "OW"),
  MTmethod = c("HIMA", "HIMA2"),
  tips = TRUE,
  simplify = TRUE
)

Arguments

X

Binary exposure.

M

High dimensional mediators.

Y

Continuous outcome.

COV

Covariates, also called confounders.

k

SIS parameter.

CONFmethod

Method of confounding adjustments, including RA,PS,IPW,OW.

MTmethod

Method of mediation testing, including HIMA, HIMA2.

tips

Whether to output message.

simplify

Only output mediator which p value less than 0.05.

Value

A dataframe contain mediators' point estimate and p value.

Examples

library(mydemo)
dat <- sim_data(n = 300, p = 2000)
X <- dat$X
Y <- dat$Y
M <- dat$M
COV <- dat$COV
himaconf(X, M, Y, COV, CONFmethod = "OW", MTmethod = "HIMA2")
#> Step 1: Sure Independent Screening ...  (2025-02-06 00:51:45.937636)
#> Step 2: DLasso Penalized Estimation ...   (2025-02-06 00:51:49.668616)
#> Nodewise regressions will be computed as no argument Z was provided.
#> You can store Z to avoid the majority of the computation next time around.
#> Z only depends on the design matrix x.
#> Step 3: Mediation test ...     (2025-02-06 00:52:52.806649)
#> Done!     (2025-02-06 00:52:53.03491)
#>          alpha  alpha_se      alpha_p      beta    beta_se       beta_p
#> M3   0.3401660 0.1468601 0.0212224199 0.5021577 0.04961240 4.430372e-24
#> M4   0.6094667 0.1547605 0.0001023138 0.5712314 0.04746269 2.315684e-33
#> M320 0.3569214 0.1574147 0.0240821511 0.1259919 0.04839477 9.229888e-03
#>             IDE        P_fdr       P_test  Med_prop        DE     DE_se
#> M3   0.17081699 2.049157e-02 0.0212224199 12.435713 0.6739322 0.1251152
#> M4   0.34814653 1.109621e-06 0.0001023138 25.345549 0.6739322 0.1251152
#> M320 0.04496921 2.049157e-02 0.0240821511  3.273821 0.6739322 0.1251152
#>              DE_p     TE
#> M3   7.184524e-08 1.3736
#> M4   7.184524e-08 1.3736
#> M320 7.184524e-08 1.3736